+++ app/brochure/routes/index.js
@@ -106,11 +106,10 @@ brochure.param("subsubsection", function(req, res, next, subsubsection) {
   next();
 });
 
-brochure.get("/", function(req, res) {
+brochure.get("/", require("./featured"), function(req, res) {
   res.locals.title = "Blot – A blogging platform with no interface";
   res.locals.selected.index = "selected";
-  // Only show the first six sites on the homepage
-  res.locals.featured = require("./featured").slice(0, 6);
+  res.locals.featured = res.locals.featured.slice(0, 6);
   res.render("index");
 });
